refactor repetitive code
