+++ scripts/plot_multiple_precision_vs_cache_hit_ratio.py
@@ -1,64 +1,16 @@
 import argparse
 import os
-import re
 
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
-
-
-def get_model_name_from_dir(dir_name):
-    name = dir_name
-    if name.startswith("neural_"):
-        name = name[7:]
-    known_prefixes = ["redis", "Alibaba-NLP"]
-    for prefix in known_prefixes:
-        if name.startswith(prefix + "_"):
-            return name.replace(prefix + "_", prefix + "/", 1)
-    match = re.search(r"_\d{4}_\d{2}_\d{2}_\d{2}_\d{2}_\d{2}", name)
-    if match:
-        name = name[: match.start()]
-        for prefix in known_prefixes:
-            if name.startswith(prefix + "_"):
-                return name.replace(prefix + "_", prefix + "/", 1)
-    return name
-
-
-def crawl_results(base_dir):
-    results = {}
-    if not os.path.exists(base_dir):
-        return results
-    for dataset in os.listdir(base_dir):
-        dataset_path = os.path.join(base_dir, dataset)
-        if not os.path.isdir(dataset_path):
-            continue
-        results[dataset] = {}
-        for model_dir in os.listdir(dataset_path):
-            model_path = os.path.join(dataset_path, model_dir)
-            if not os.path.isdir(model_path):
-                continue
-            if any(sub.startswith("run_") for sub in os.listdir(model_path)):
-                model_name = get_model_name_from_dir(model_dir)
-                if model_name not in results[dataset]:
-                    results[dataset][model_name] = []
-                for run_dir in os.listdir(model_path):
-                    if not run_dir.startswith("run_"):
-                        continue
-                    run_path = os.path.join(model_path, run_dir)
-                    timestamp_dirs = [d for d in os.listdir(run_path) if os.path.isdir(os.path.join(run_path, d))]
-                    if timestamp_dirs:
-                        timestamp_dirs.sort()
-                        results[dataset][model_name].append(os.path.join(run_path, timestamp_dirs[-1]))
-            elif "threshold_sweep_results.csv" in os.listdir(model_path):
-                model_name = get_model_name_from_dir(model_dir)
-                if model_name not in results[dataset]:
-                    results[dataset][model_name] = []
-                results[dataset][model_name].append(model_path)
-    return results
+from utils import crawl_results
 
 
 def main():
-    parser = argparse.ArgumentParser("Usage: python plot_multiple_precision_vs_cache_hit_ratio.py --base_dir <base_dir>")
+    parser = argparse.ArgumentParser(
+        "Usage: python plot_multiple_precision_vs_cache_hit_ratio.py --base_dir <base_dir>"
+    )
     parser.add_argument("--base_dir", type=str, default="complete_benchmark_results")
     args = parser.parse_args()
     base_dir = args.base_dir
